Fixed MAGE-5496: Default value that was specified in system settings does not presented in Code Format drop-down on Manage Coupon Codes tab

+++ app/code/core/Mage/SalesRule/Helper/Coupon.php
@@ -91,7 +91,7 @@ class Mage_SalesRule_Helper_Coupon extends Mage_Core_Helper_Abstract
      */
     public function getDefaultFormat()
     {
-        return (int)Mage::getStoreConfig(self::XML_PATH_SALES_RULE_COUPON_FORMAT);
+        return Mage::getStoreConfig(self::XML_PATH_SALES_RULE_COUPON_FORMAT);
     }
 
     /**
